+++ src/VaultSharp/V1/SecretsEngines/Active Directory/ActiveDirectorySecretsEngineProvider.cs
@@ -16,6 +16,9 @@ namespace VaultSharp.V1.SecretsEngines.ActiveDirectory
 
         public async Task<Secret<ActiveDirectoryCredentials>> GetCredentialsAsync(string roleName, string mountPoint = SecretsEngineDefaultPaths.ActiveDirectory, string wrapTimeToLive = null)
         {
+            Checker.NotNull(mountPoint, "mountPoint");
+            Checker.NotNull(roleName, "roleName");
+
             return await _polymath.MakeVaultApiRequest<Secret<ActiveDirectoryCredentials>>("v1/" + mountPoint.Trim('/') + "/creds/" + roleName.Trim('/'), HttpMethod.Get, wrapTimeToLive: wrapTimeToLive).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
         }
     }